Downgrading output into floats, after some testing double precision wasn't necessary
git-svn-id: http://jopensurf.googlecode.com/svn/trunk@14 714e5f76-d679-94a1-201b-066bf2ce06a2

+++ src/com/stromberglabs/jopensurf/Clusterable.java
@@ -1,5 +1,5 @@
 package com.stromberglabs.jopensurf;
 
 public interface Clusterable {
-	public double[] getLocation();
+	public float[] getLocation();
 }
